VisaGuard Frontend Design/univisa-backend/requirements.txt
# UniVisa Backend - Python 3.11+
# Core API only. Numpy first + wheel-only so Windows doesn't try to build from source.
numpy>=2.0.0

fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# AI & Embeddings
anthropic==0.39.0
sentence-transformers==3.3.1

# Vector store (RAG) - ChromaDB for hackathon
chromadb==0.5.23

# Optional: for scripts/fetch_reddit.py (pure Python)
praw==7.7.1
